requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
absl-py=2.0.0=pyhd8ed1ab_0
appnope=0.1.3=pyhd8ed1ab_0
asgiref=3.7.2=pypi_0
asttokens=2.4.0=pyhd8ed1ab_0
astunparse=1.6.3=pyhd8ed1ab_0
asyncio=3.4.3=pypi_0
attrs=23.1.0=pypi_0
autobahn=23.6.2=pypi_0
automat=22.10.0=pypi_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=pyhd8ed1ab_3
backports.functools_lru_cache=1.6.5=pyhd8ed1ab_0
blinker=1.7.0=pyhd8ed1ab_0
brotli=1.1.0=hb547adb_1
brotli-bin=1.1.0=hb547adb_1
brotli-python=1.1.0=py310h1253130_1
bzip2=1.0.8=h3422bc3_4
c-ares=1.21.0=h93a5062_0
ca-certificates=2023.7.22=hf0a4a13_0
cached-property=1.5.2=hd8ed1ab_1
cached_property=1.5.2=pyha770c72_1
cachetools=4.2.4=pyhd8ed1ab_0
certifi=2023.7.22=pyhd8ed1ab_0
cffi=1.16.0=py310hdcd7c05_0
channels=4.0.0=pypi_0
charset-normalizer=3.3.0=pyhd8ed1ab_0
click=8.1.7=unix_pyh707e725_0
comm=0.1.4=pyhd8ed1ab_0
constantly=23.10.4=pypi_0
contourpy=1.1.1=py310h38f39d4_1
cryptography=41.0.5=py310h0e6f4b3_0
cycler=0.12.1=pyhd8ed1ab_0
daphne=4.0.0=pypi_0
debugpy=1.8.0=py310h1253130_1
decorator=5.1.1=pyhd8ed1ab_0
django=4.2.6=pypi_0
exceptiongroup=1.1.3=pyhd8ed1ab_0
executing=1.2.0=pyhd8ed1ab_0
filelock=3.12.4=pyhd8ed1ab_0
flatbuffers=23.5.26=h13dd4ca_1
fonttools=4.43.1=py310hd125d64_0
freetype=2.12.1=hadb7bae_2
gast=0.5.4=pyhd8ed1ab_0
giflib=5.2.1=h1a8c8d9_3
gmp=6.2.1=h9f76cd9_0
gmpy2=2.1.2=py310h2e6cad2_1
google-auth=1.21.3=py_0
google-auth-oauthlib=1.0.0=pyhd8ed1ab_0
google-pasta=0.2.0=pyh8c360ce_0
grpcio=1.54.3=py310h1253130_0
gunicorn=21.2.0=pypi_0
h11=0.14.0=pypi_0
h5py=3.10.0=nompi_py310h428b110_100
hdf5=1.14.2=nompi_h3aba7b3_100
hyperlink=21.0.0=pypi_0
icu=73.2=hc8870d7_0
idna=3.4=pyhd8ed1ab_0
importlib-metadata=6.8.0=pyha770c72_0
importlib_metadata=6.8.0=hd8ed1ab_0
incremental=22.10.0=pypi_0
ipykernel=6.25.2=pyh1050b4e_0
ipython=8.16.1=pyh31c8845_0
jedi=0.19.1=pyhd8ed1ab_0
jinja2=3.1.2=pyhd8ed1ab_1
jupyter_client=8.4.0=pyhd8ed1ab_0
jupyter_core=5.4.0=py310hbe9552e_0
keras=2.13.1=pyhd8ed1ab_0
kiwisolver=1.4.5=py310h38f39d4_1
krb5=1.21.2=h92f50d5_0
lcms2=2.15=hf2736f0_3
lerc=4.0.0=h9a09cb3_0
libabseil=20230125.3=cxx17_h13dd4ca_0
libaec=1.1.2=h13dd4ca_1
libblas=3.9.0=19_osxarm64_openblas
libbrotlicommon=1.1.0=hb547adb_1
libbrotlidec=1.1.0=hb547adb_1
libbrotlienc=1.1.0=hb547adb_1
libcblas=3.9.0=19_osxarm64_openblas
libcurl=8.4.0=h2d989ff_0
libcxx=16.0.6=h4653b0c_0
libdeflate=1.19=hb547adb_0
libedit=3.1.20191231=hc8eb9b7_2
libev=4.33=h642e427_1
libffi=3.4.2=h3422bc3_5
libgfortran=5.0.0=13_2_0_hd922786_1
libgfortran5=13.2.0=hf226fd6_1
libgrpc=1.54.3=h0a338ca_0
libjpeg-turbo=3.0.0=hb547adb_1
liblapack=3.9.0=19_osxarm64_openblas
libnghttp2=1.58.0=ha4dd798_0
libopenblas=0.3.24=openmp_hd76b1f2_0
libpng=1.6.39=h76d750c_0
libprotobuf=3.21.12=ha614eb4_2
libsodium=1.0.18=h27ca646_1
libsqlite=3.43.2=h091b4b1_0
libssh2=1.11.0=h7a5bd25_0
libtiff=4.6.0=ha8a6c65_2
libuv=1.46.0=hb547adb_0
libwebp-base=1.3.2=hb547adb_0
libxcb=1.15=hf346824_0
libzlib=1.2.13=h53f4e23_5
llvm-openmp=17.0.2=h1c12783_0
markdown=3.5.1=pyhd8ed1ab_0
markupsafe=2.1.3=py310h2aa6e3c_1
matplotlib=3.8.0=py310hb6292c7_2
matplotlib-base=3.8.0=py310h9d2df84_2
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mpc=1.3.1=h91ba8db_0
mpfr=4.2.0=he09a6ba_0
mpmath=1.3.0=pyhd8ed1ab_0
munkres=1.1.4=pyh9f0ad1d_0
ncurses=6.4=h7ea286d_0
nest-asyncio=1.5.8=pyhd8ed1ab_0
networkx=3.1=pyhd8ed1ab_0
nomkl=1.0=h5ca1d4c_0
numpy=1.26.0=py310he2cad68_0
oauthlib=3.2.2=pyhd8ed1ab_0
openjpeg=2.5.0=h4c1507b_3
openssl=3.1.4=h0d3ecfb_0
opt_einsum=3.3.0=pyhc1e730c_2
packaging=23.2=pyhd8ed1ab_0
pandas=2.1.1=py310h5924a0a_1
parso=0.8.3=pyhd8ed1ab_0
patsy=0.5.3=pyhd8ed1ab_0
pexpect=4.8.0=pyh1a96a4e_2
pickleshare=0.7.5=py_1003
pillow=10.1.0=py310hfae7ebd_0
pip=23.3=pyhd8ed1ab_0
platformdirs=2.6.0=pyhd8ed1ab_0
prompt-toolkit=3.0.39=pyha770c72_0
prompt_toolkit=3.0.39=hd8ed1ab_0
protobuf=4.21.12=py310h0f1eb42_0
psutil=5.9.5=py310h2aa6e3c_1
pthread-stubs=0.4=h27ca646_1001
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
pyasn1=0.5.0=pyhd8ed1ab_0
pyasn1-modules=0.3.0=pyhd8ed1ab_0
pycparser=2.21=pyhd8ed1ab_0
pygments=2.16.1=pyhd8ed1ab_0
pyjwt=2.8.0=pyhd8ed1ab_0
pyopenssl=23.3.0=pypi_0
pyparsing=3.1.1=pyhd8ed1ab_0
pysocks=1.7.1=pyha2e5f31_6
python=3.10.12=h01493a6_0_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python-flatbuffers=23.5.26=pyhd8ed1ab_0
python-tzdata=2023.3=pyhd8ed1ab_0
python_abi=3.10=4_cp310
pytorch=2.0.0=cpu_generic_py310h3496f23_1
pytz=2023.3.post1=pyhd8ed1ab_0
pyzmq=25.1.1=py310h30b7201_1
re2=2023.03.02=hc5e2d97_0
readline=8.2=h92ec313_1
requests=2.31.0=pyhd8ed1ab_0
requests-oauthlib=1.3.1=pyhd8ed1ab_0
rsa=4.9=pyhd8ed1ab_0
scipy=1.11.3=py310hd1cfc7d_1
seaborn=0.13.0=hd8ed1ab_0
seaborn-base=0.13.0=pyhd8ed1ab_0
service-identity=23.1.0=pypi_0
setuptools=68.2.2=pyhd8ed1ab_0
six=1.16.0=pyh6c4a22f_0
sleef=3.5.1=h156473d_2
snappy=1.1.10=h17c5cce_0
sqlparse=0.4.4=pypi_0
stack_data=0.6.2=pyhd8ed1ab_0
statsmodels=0.14.0=py310h50ce23c_2
sympy=1.12=pypyh9d50eac_103
tensorboard=2.13.0=pyhd8ed1ab_0
tensorboard-data-server=0.7.0=py310hdd3b5e7_1
tensorflow=2.13.1=cpu_py310h6ecea76_1
tensorflow-base=2.13.1=cpu_py310h0e84cb2_1
tensorflow-estimator=2.13.1=cpu_py310h0d828bf_1
termcolor=2.3.0=pyhd8ed1ab_0
tflearn=0.5.0=pypi_0
tk=8.6.13=hb31c410_0
torchvision=0.15.2=cpu_py310h958d273_2
tornado=6.3.3=py310h2aa6e3c_1
traitlets=5.11.2=pyhd8ed1ab_0
twisted=23.10.0=pypi_0
txaio=23.1.1=pypi_0
typing_extensions=4.5.0=pyha770c72_0
tzdata=2023c=h71feb2d_0
unicodedata2=15.1.0=py310h2aa6e3c_0
urllib3=2.0.6=pyhd8ed1ab_0
uvicorn=0.23.2=pypi_0
wcwidth=0.2.8=pyhd8ed1ab_0
werkzeug=3.0.1=pyhd8ed1ab_0
wheel=0.41.2=pyhd8ed1ab_0
wrapt=1.15.0=py310h2aa6e3c_1
xorg-libxau=1.0.11=hb547adb_0
xorg-libxdmcp=1.1.3=h27ca646_0
xz=5.2.6=h57fd34a_0
zeromq=4.3.5=h965bd2d_0
zipp=3.17.0=pyhd8ed1ab_0
zlib=1.2.13=h53f4e23_5
zope-interface=6.1=pypi_0
zstd=1.5.5=h4f39d0f_0
